Now possible to specify omc for exec
The user of the OMJulia API can now use both OMCSession() and OMCSession(path_to_executable) and it will work as before.

This PL does not add this support for Windows. It will still only execute the old function. However, it will prompt the user of incorrect usage. I wrote an assert macro for that